work-tree/work_tree/requirements.txt
asgiref==3.4.1
bootstrap4==0.1.0
Django==3.2.9
django-ckeditor==6.2.0
django-crispy-forms==1.13.0
django-js-asset==1.2.2
django-modeltranslation==0.17.1
django-mptt==0.12.0
humanize==3.5.0
Pillow==8.4.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
